feat: refactor trade handling to use Quantity value object and update response types

+++ src/Scheduled.Stock.Purchase.Api/Features/Trades/CreateTrade/CreateTradeRequest.cs
@@ -3,7 +3,7 @@ namespace Scheduled.Stock.Purchase.Api.Features.Trades.CreateTrade;
 internal sealed class CreateTradeRequest
 {
     public string Ticker { get; init; } = default!;
-    public decimal Quantity { get; init; }
+    public int Quantity { get; init; }
     public decimal Price { get; init; }
     public string Type { get; init; } = default!;
 }
